backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
mysqlclient==2.2.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
openai==1.3.5
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.5
prometheus-client==0.19.0
python-dotenv==1.0.0
httpx==0.25.2